ocr-service/requirements.txt
pycountry==18.5.26
pymongo==3.7.1
Flask==1.0.2
Flask-Cors==3.0.6
Jinja2==2.10
MarkupSafe==1.0
Werkzeug==0.14.1
asn1crypto==0.24.0
azure-common==1.1.14
azure-nspkg==2.0.0
azure-storage==0.36.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.3
gunicorn==19.9.0
idna==2.7
itsdangerous==0.24
pip==9.0.1
pycparser==2.18
python-dateutil==2.7.3
requests==2.19.1
setuptools==28.8.0
six==1.11.0
urllib3==1.23